Add customer account type
This commit adds a new account type called "customer" to the `AccountType` enum in the `User.swift` file. The `customer` account type is used to represent customers in the system.

It also updates the `Key.swift` file to change the schema name to "crypto_keys" and sets the database ID to `.keyManagement`. This change improves the organization and management of cryptographic keys.

Additionally, it introduces a new file called `CategoryDTO.swift` in the `DTOs/Products` directory. This file defines a `CategoryDTO` struct that conforms to the `Content` protocol. The `CategoryDTO` struct represents a category of products and includes properties for the category's ID and name. It also provides methods to convert the DTO to a model object.

Finally, the `CertificateController.swift` file is modified to include a new method called `fetchPublicKey`. This method retrieves the public key from the database using the `.keyManagement` database ID.

These changes enhance the functionality and organization of the application.