Add FileError enum for handling file saving errors
This commit adds a new enum called `FileError` in the `FileUploads` directory. The `FileError` enum is used to represent errors that occur during file saving. It includes a case for `couldNotSave` with a reason string parameter. This enum will be used to handle file saving errors in the application.

Also, a new enum called `StorageType` is added in the `FileUploads` directory. This enum represents different storage types for files and includes cases for `local` and `s3`. It conforms to the `Codable` and `Sendable` protocols.

Additionally, the `OrderControllerV1.swift`, `TransactionControllerV1.swift`, and `AuthenicationControllerV1.swift` files are renamed to `OrderController.swift`, `TransactionController.swift`, and `AuthenicationController.swift` respectively. This change improves the naming consistency of the controller files.

Lastly, the `Package.resolved` file and `Package.swift` file are modified to update dependencies and remove unused dependencies.

These changes enhance the file handling capabilities and improve the organization of the codebase.